requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
openai>=1.3.0
python-dotenv>=1.0.0
numpy>=1.21.0
requests>=2.25.0
werkzeug>=2.3.0
